\name{bonuslabgroup8-package}
\alias{bonuslabgroup8-package}
\alias{bonuslabgroup8}
\docType{package}
\title{
\packageTitle{bonuslabgroup8}
}
\description{
\packageDescription{bonuslabgroup8}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bonuslabgroup8}
\packageIndices{bonuslabgroup8}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{bonuslabgroup8}

Maintainer: \packageMaintainer{bonuslabgroup8}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
