% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression}
\arguments{
\item{formula}{Takes a formula of the form Y~X.}

\item{data}{Takes a dataset in the form of a data.frame.}

\item{lambda}{Is the constraint on the parameters.}
}
\value{
Returns an object of the class 'ridgereg'. This object can be manipulated.
}
\description{
Running a Ridge Reregression Model.
}

