\name{bonuslabgroup8-package}
\alias{bonuslabgroup8-package}
\alias{bonuslabgroup8}
\docType{package}
\title{
\packageTitle{bonuslabgroup8}
}
\description{
\packageDescription{bonuslabgroup8}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bonuslabgroup8}
\packageIndices{bonuslabgroup8}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{bonuslabgroup8}

Maintainer: \packageMaintainer{bonuslabgroup8}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }

